hello-spring-review/src/main/java/hello/hellospringreview/service/MemberService.java
package hello.hellospringreview.service;

import hello.hellospringreview.domain.Member;
import hello.hellospringreview.repository.MemberRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import java.util.List;
import java.util.Optional;

@Transactional
public class MemberService {

    private MemberRepository memberRepository;

    public MemberService(MemberRepository memberRepository){
        this.memberRepository = memberRepository;
    }

    public Long join(Member member){
        validateDuplicateMember(member);
        memberRepository.save(member);
        return member.getId();
    }

    public void validateDuplicateMember(Member member){
        memberRepository.findByName(member.getName()).ifPresent(m -> {
            throw new IllegalStateException("이미 존재하는 회원입니다.");
        });
    }

    public Optional<Member> findOne(Long id){
        return memberRepository.findById(id);
    }

    public List<Member> findMembers(){
        return memberRepository.findAll();
    }

}
